Added edits in the module4-task1 (Edited arrays names and edited condition in the createAuthor function). Made tasks for module5-task1 (Created JS modules)
